safari-600.7-branch/LayoutTests/fast/dom/setAttributeNS.html

<!DOCTYPE html>
<script src="../../resources/js-test-pre.js"></script>

<script>
description("Test that setAttributeNS with a null namespaceURI actually sets it to the value null and not a string with the contents 'null'.");

window.elem = document.createElementNS('http://www.example.org', 'test');
elem.setAttributeNS(null, 'name', 'value');
shouldBeNull('elem.attributes[0].namespaceURI');
</script>

<script src="../../resources/js-test-post.js"></script>
